Upgrade Java version to 21

+++ app/src/main/java/freemind/view/mindmapview/MapView.java
@@ -249,8 +249,8 @@ public class MapView extends JPanel implements ViewAbstraction, Printable, Autos
 	private Rectangle boundingRectangle = null;
 	private boolean fitToPage = true;
 
-	int mPaintingTime;
-	int mPaintingAmount;
+	long mPaintingTime;
+	long mPaintingAmount;
 	static boolean printOnWhiteBackground;
 	static Color standardMapBackgroundColor;
 	static Color standardSelectColor;
